+++ packages/terra-notification-dialog/src/terra-dev-site/doc/example/ActionOrderExample.jsx
@@ -22,7 +22,7 @@ const NoVariantNotificationDialog = () => {
         buttonOrder={buttonOrder}
         isOpen={isOpen}
         variant="info"
-        title="Use a title that relates directly to the choices"
+        dialogTitle="Use a title that relates directly to the choices"
         startMessage="The Main Instruction is text used to provide more detail or define terminology. Don’t repeat the title verbatim."
         acceptAction={{
           text: 'accept',