+++ packages/terra-notification-dialog/src/terra-dev-site/doc/example/CustomVariantExample.jsx
@@ -28,7 +28,7 @@ const NotificationDialogExample = () => {
           signalWord: 'Custom Dialog Signal Word',
         }}
         isOpen={isOpen}
-        title="Use a title that relates directly to the choices"
+        dialogTitle="Use a title that relates directly to the choices"
         startMessage="The Main Instruction is text used to provide more detail or define terminology. Don’t repeat the title verbatim."
         acceptAction={{
           text: 'accept',