+++ packages/terra-notification-dialog/src/terra-dev-site/doc/example/EmphasizedRejectExample.jsx
@@ -18,7 +18,7 @@ const ReversedActionNotificationDialog = () => {
       <NotificationDialog
         variant="alert"
         isOpen={isOpen}
-        title="Use a title that relates directly to the choices"
+        dialogTitle="Use a title that relates directly to the choices"
         startMessage="The Main Instruction is text used to provide more detail or define terminology. Don’t repeat the title verbatim."
         acceptAction={{
           text: 'Confirm',