+++ packages/terra-notification-dialog/src/terra-dev-site/doc/example/InfoVariantExample.jsx
@@ -19,7 +19,7 @@ const NotificationDialogExample = () => {
       <NotificationDialog
         variant="info"
         isOpen={isOpen}
-        title="Use a title that relates directly to the choices"
+        dialogTitle="Use a title that relates directly to the choices"
         startMessage="The Main Instruction is text used to provide more detail or define terminology. Don’t repeat the title verbatim."
         acceptAction={{
           text: 'accept',