+++ packages/terra-notification-dialog/src/terra-dev-site/test/notification-dialog/NotificationDialogVariant.jsx
@@ -36,7 +36,7 @@ const NotificationDialogVariant = (props) => {
         {...customProps}
         variant={variant}
         isOpen={isOpen}
-        title="The title relates directly to the choices."
+        dialogTitle="The title relates directly to the choices."
         startMessage="The Main Instruction is text used to provide more detail or define terminology. Don’t repeat the title verbatim."
         acceptAction={{
           text: 'Confirm',