+++ packages/terra-notification-dialog/src/terra-dev-site/test/notification-dialog/RejectFirstNotificationDialog.test.jsx
@@ -22,7 +22,7 @@ const RejectFirstNotificationDialog = () => {
       <NotificationDialog
         variant="alert"
         isOpen={isOpen}
-        title="Make sure that the title relates directly to the choices."
+        dialogTitle="Make sure that the title relates directly to the choices."
         startMessage="The Main Instruction is text used to provide more detail or define terminology. Don’t repeat the title verbatim."
         acceptAction={{
           text: 'accept',